+++ test/byContract/IexecPocoBoost/IexecPocoBoostDelegate.test.ts
@@ -32,8 +32,9 @@ import {
     signOrder,
     hashOrder,
     signOrders,
-    Iexec,
     IexecAccounts,
+    OrdersAssets,
+    OrdersPrices,
 } from '../../../utils/createOrders';
 import {
     buildAndSignSchedulerMessage,
@@ -49,7 +50,6 @@ chai.use(smock.matchers);
 
 // TODO: Rename to teeDealTag
 const dealTagTee = '0x0000000000000000000000000000000000000000000000000000000000000001';
-const standardDealTag = '0x0000000000000000000000000000000000000000000000000000000000000000';
 const taskIndex = 0;
 const volume = taskIndex + 1;
 const schedulerRewardRatio = 1;
@@ -59,6 +59,9 @@ const BALANCES = 'm_balances';
 const FROZENS = 'm_frozens';
 const WORKERPOOL_STAKE_RATIO = 30;
 const { domain, domainSeparator } = buildDomain();
+const appPrice = 1000;
+const datasetPrice = 1_000_000;
+const workerpoolPrice = 1_000_000_000;
 
 async function deployBoostFixture() {
     const [
@@ -138,7 +141,8 @@ describe('IexecPocoBoostDelegate', function () {
     let [appProvider, datasetProvider, scheduler, worker, enclave, requester, beneficiary, anyone] =
         [] as SignerWithAddress[];
     let accounts: IexecAccounts;
-    let entriesAndRequester: Iexec<string>;
+    let ordersAssets: OrdersAssets;
+    let ordersPrices: OrdersPrices;
 
     beforeEach('set up contract instances and mock app', async () => {
         const fixtures = await loadFixture(deployBoostFixture);
@@ -156,15 +160,20 @@ describe('IexecPocoBoostDelegate', function () {
             dataset: datasetProvider,
             workerpool: scheduler,
             requester: requester,
+            beneficiary: beneficiary,
         };
         appInstance = await createMock<App__factory, App>('App');
         workerpoolInstance = await createMock<Workerpool__factory, Workerpool>('Workerpool');
         datasetInstance = await createMock<Dataset__factory, Dataset>('Dataset');
-        entriesAndRequester = {
+        ordersAssets = {
             app: appInstance.address,
             dataset: datasetInstance.address,
             workerpool: workerpoolInstance.address,
-            requester: requester.address,
+        };
+        ordersPrices = {
+            app: appPrice,
+            dataset: datasetPrice,
+            workerpool: workerpoolPrice,
         };
         appRegistry = await smock.fake<AppRegistry>(AppRegistry__factory);
         await iexecPocoBoostInstance.setVariable('m_appregistry', appRegistry.address);
@@ -182,7 +191,7 @@ describe('IexecPocoBoostDelegate', function () {
     });
 
     describe('Match Orders Boost', function () {
-        it('Should match orders', async function () {
+        it('Should match orders (TEE)', async function () {
             appInstance.owner.returns(appProvider.address);
             workerpoolInstance.owner.returns(scheduler.address);
             datasetInstance.owner.returns(datasetProvider.address);
@@ -192,16 +201,14 @@ describe('IexecPocoBoostDelegate', function () {
             const datasetPrice = 1_000_000;
             const workerpoolPrice = 1_000_000_000;
             const { orders, appOrder, datasetOrder, workerpoolOrder, requestOrder } =
-                buildCompatibleOrders(entriesAndRequester, dealTagTee, {
-                    app: appPrice,
-                    dataset: datasetPrice,
-                    workerpool: workerpoolPrice,
+                buildCompatibleOrders({
+                    assets: ordersAssets,
+                    requester: requester.address,
+                    beneficiary: beneficiary.address,
+                    tag: dealTagTee,
+                    prices: ordersPrices,
+                    callback: ethers.Wallet.createRandom().address,
                 });
-            requestOrder.requester = requester.address;
-            requestOrder.beneficiary = beneficiary.address;
-
-            // Set callback
-            requestOrder.callback = ethers.Wallet.createRandom().address;
             // Should match orders with low app order volume
             // Set volumes
             appOrder.volume = 2; // smallest unconsumed volume among all orders
@@ -273,6 +280,7 @@ describe('IexecPocoBoostDelegate', function () {
                     requestOrder.category,
                     dealTagTee,
                     requestOrder.params,
+                    beneficiary.address,
                 )
                 .to.emit(iexecPocoBoostInstance, 'OrdersMatched')
                 .withArgs(
@@ -308,13 +316,14 @@ describe('IexecPocoBoostDelegate', function () {
                     (100 - schedulerRewardRatio)) / // worker ratio
                     100,
             );
-            expect(deal.beneficiary).to.be.equal(requestOrder.beneficiary, 'Beneficiary mismatch');
             expect(deal.deadline).to.be.equal(
                 startTime + // match order block timestamp
                     7 * // contribution deadline ratio
                         60, // requested category time reference
             );
-            expect(deal.callback).to.be.equal(requestOrder.callback, 'Callback mismatch');
+            expect(deal.callback)
+                .to.be.equal(requestOrder.callback, 'Callback mismatch')
+                .to.not.be.equal(constants.NULL.ADDRESS);
             // Check prices.
             expect(deal.workerpoolPrice).to.be.equal(
                 workerpoolOrder.workerpoolprice,
@@ -348,13 +357,17 @@ describe('IexecPocoBoostDelegate', function () {
             );
         });
 
-        it('Should match orders with pre-signatures', async function () {
+        it('Should match orders with pre-signatures (TEE)', async function () {
             appInstance.owner.returns(appProvider.address);
             workerpoolInstance.owner.returns(scheduler.address);
             datasetInstance.owner.returns(datasetProvider.address);
             const { appOrder, datasetOrder, workerpoolOrder, requestOrder } = buildCompatibleOrders(
-                entriesAndRequester,
-                dealTagTee,
+                {
+                    assets: ordersAssets,
+                    requester: requester.address,
+                    beneficiary: beneficiary.address,
+                    tag: dealTagTee,
+                },
             );
             const appOrderHash = hashOrder(domain, appOrder);
             const datasetOrderHash = hashOrder(domain, datasetOrder);
@@ -385,6 +398,7 @@ describe('IexecPocoBoostDelegate', function () {
                     requestOrder.category,
                     dealTagTee,
                     requestOrder.params,
+                    beneficiary.address,
                 )
                 .to.emit(iexecPocoBoostInstance, 'OrdersMatched')
                 .withArgs(
@@ -397,7 +411,7 @@ describe('IexecPocoBoostDelegate', function () {
                 );
         });
 
-        it('Should match orders without dataset', async function () {
+        it('Should match orders without dataset (TEE)', async function () {
             appInstance.owner.returns(appProvider.address);
             workerpoolInstance.owner.returns(scheduler.address);
 
@@ -410,30 +424,27 @@ describe('IexecPocoBoostDelegate', function () {
 
             const { appOrder, datasetOrder, workerpoolOrder, requestOrder } = buildCompatibleOrders(
                 {
-                    app: appInstance.address,
-                    dataset: constants.NULL.ADDRESS, // No dataset.
-                    workerpool: workerpoolInstance.address,
+                    assets: {
+                        app: appInstance.address,
+                        dataset: constants.NULL.ADDRESS, // No dataset.
+                        workerpool: workerpoolInstance.address,
+                    },
                     requester: requester.address,
-                },
-                dealTagTee,
-                {
-                    app: appPrice,
-                    dataset: 0,
-                    workerpool: workerpoolPrice,
+                    beneficiary: beneficiary.address,
+                    prices: {
+                        app: appPrice,
+                        dataset: 0,
+                        workerpool: workerpoolPrice,
+                    },
+                    tag: dealTagTee,
                 },
             );
-
-            requestOrder.requester = requester.address;
-            requestOrder.beneficiary = beneficiary.address;
-
-            // Set callback
-            requestOrder.callback = ethers.Wallet.createRandom().address;
             await signOrder(domain, appOrder, appProvider);
             await signOrder(domain, workerpoolOrder, scheduler);
             await signOrder(domain, requestOrder, requester);
             const dealId = getDealId(domain, requestOrder, taskIndex);
-            const datasetOrderHash = hashOrder(domain, datasetOrder);
-            await expectOrderConsumed(iexecPocoBoostInstance, datasetOrderHash, undefined);
+            const emptyDatasetOrderHash = hashOrder(domain, datasetOrder);
+            await expectOrderConsumed(iexecPocoBoostInstance, emptyDatasetOrderHash, undefined);
 
             await expect(
                 iexecPocoBoostInstance.matchOrdersBoost(
@@ -452,6 +463,7 @@ describe('IexecPocoBoostDelegate', function () {
                     requestOrder.category,
                     dealTagTee,
                     requestOrder.params,
+                    beneficiary.address,
                 )
                 .to.emit(iexecPocoBoostInstance, 'OrdersMatched')
                 .withArgs(
@@ -462,7 +474,7 @@ describe('IexecPocoBoostDelegate', function () {
                     hashOrder(domain, requestOrder),
                     volume,
                 );
-            await expectOrderConsumed(iexecPocoBoostInstance, datasetOrderHash, undefined);
+            await expectOrderConsumed(iexecPocoBoostInstance, emptyDatasetOrderHash, undefined);
             const deal = await viewDealBoost(dealId);
             expect(deal.datasetPrice).to.be.equal(0);
         });
@@ -472,7 +484,10 @@ describe('IexecPocoBoostDelegate', function () {
             workerpoolInstance.owner.returns(scheduler.address);
             datasetInstance.owner.returns(datasetProvider.address);
             const { orders, appOrder, datasetOrder, workerpoolOrder, requestOrder } =
-                buildCompatibleOrders(entriesAndRequester, dealTagTee);
+                buildCompatibleOrders({
+                    assets: ordersAssets,
+                    requester: requester.address,
+                });
             // Set volumes
             appOrder.volume = 6;
             datasetOrder.volume = 5; // smallest unconsumed volume among all orders
@@ -517,7 +532,10 @@ describe('IexecPocoBoostDelegate', function () {
             workerpoolInstance.owner.returns(scheduler.address);
             datasetInstance.owner.returns(datasetProvider.address);
             const { orders, appOrder, datasetOrder, workerpoolOrder, requestOrder } =
-                buildCompatibleOrders(entriesAndRequester, dealTagTee);
+                buildCompatibleOrders({
+                    assets: ordersAssets,
+                    requester: requester.address,
+                });
             // Set volumes
             appOrder.volume = 5;
             datasetOrder.volume = 4;
@@ -562,7 +580,10 @@ describe('IexecPocoBoostDelegate', function () {
             workerpoolInstance.owner.returns(scheduler.address);
             datasetInstance.owner.returns(datasetProvider.address);
             const { orders, appOrder, datasetOrder, workerpoolOrder, requestOrder } =
-                buildCompatibleOrders(entriesAndRequester, dealTagTee);
+                buildCompatibleOrders({
+                    assets: ordersAssets,
+                    requester: requester.address,
+                });
             // Set volumes
             appOrder.volume = 7;
             datasetOrder.volume = 6;
@@ -607,7 +628,10 @@ describe('IexecPocoBoostDelegate', function () {
             workerpoolInstance.owner.returns(scheduler.address);
             datasetInstance.owner.returns(datasetProvider.address);
             const { orders, appOrder, datasetOrder, workerpoolOrder, requestOrder } =
-                buildCompatibleOrders(entriesAndRequester, dealTagTee);
+                buildCompatibleOrders({
+                    assets: ordersAssets,
+                    requester: requester.address,
+                });
             appOrder.volume = 8;
             datasetOrder.volume = 8;
             requestOrder.volume = 8;
@@ -677,7 +701,10 @@ describe('IexecPocoBoostDelegate', function () {
             workerpoolInstance.owner.returns(scheduler.address);
             datasetInstance.owner.returns(datasetProvider.address);
             const { orders, appOrder, datasetOrder, workerpoolOrder, requestOrder } =
-                buildCompatibleOrders(entriesAndRequester, dealTagTee);
+                buildCompatibleOrders({
+                    assets: ordersAssets,
+                    requester: requester.address,
+                });
             // Set volumes
             appOrder.volume = 0; // nothing to consume
             await signOrders(domain, orders, accounts);
@@ -694,8 +721,10 @@ describe('IexecPocoBoostDelegate', function () {
 
         it('Should fail when trust is not zero', async function () {
             const { appOrder, datasetOrder, workerpoolOrder, requestOrder } = buildCompatibleOrders(
-                entriesAndRequester,
-                dealTagTee,
+                {
+                    assets: ordersAssets,
+                    requester: requester.address,
+                },
             );
             // Set non-zero trust
             requestOrder.trust = 1;
@@ -712,8 +741,10 @@ describe('IexecPocoBoostDelegate', function () {
 
         it('Should fail when categories are different', async function () {
             const { appOrder, datasetOrder, workerpoolOrder, requestOrder } = buildCompatibleOrders(
-                entriesAndRequester,
-                dealTagTee,
+                {
+                    assets: ordersAssets,
+                    requester: requester.address,
+                },
             );
             // Set different categories
             requestOrder.category = 1;
@@ -731,8 +762,10 @@ describe('IexecPocoBoostDelegate', function () {
 
         it('Should fail when category unknown', async function () {
             const { appOrder, datasetOrder, workerpoolOrder, requestOrder } = buildCompatibleOrders(
-                entriesAndRequester,
-                dealTagTee,
+                {
+                    assets: ordersAssets,
+                    requester: requester.address,
+                },
             );
             // Unknown category
             requestOrder.category = 1;
@@ -750,8 +783,10 @@ describe('IexecPocoBoostDelegate', function () {
 
         it('Should fail when app max price is less than app price', async function () {
             const { appOrder, datasetOrder, workerpoolOrder, requestOrder } = buildCompatibleOrders(
-                entriesAndRequester,
-                dealTagTee,
+                {
+                    assets: ordersAssets,
+                    requester: requester.address,
+                },
             );
             appOrder.appprice = 200;
             requestOrder.appmaxprice = 100;
@@ -768,8 +803,10 @@ describe('IexecPocoBoostDelegate', function () {
 
         it('Should fail when dataset max price is less than dataset price', async function () {
             const { appOrder, datasetOrder, workerpoolOrder, requestOrder } = buildCompatibleOrders(
-                entriesAndRequester,
-                dealTagTee,
+                {
+                    assets: ordersAssets,
+                    requester: requester.address,
+                },
             );
 
             // Set dataset price higher than dataset max price
@@ -788,8 +825,10 @@ describe('IexecPocoBoostDelegate', function () {
 
         it('Should fail when workerpool max price is less than workerpool price', async function () {
             const { appOrder, datasetOrder, workerpoolOrder, requestOrder } = buildCompatibleOrders(
-                entriesAndRequester,
-                dealTagTee,
+                {
+                    assets: ordersAssets,
+                    requester: requester.address,
+                },
             );
 
             // Set workerpool price higher than workerpool max price
@@ -809,8 +848,10 @@ describe('IexecPocoBoostDelegate', function () {
         it('Should fail when invalid app order signature', async function () {
             appInstance.owner.returns(appProvider.address);
             const { appOrder, datasetOrder, workerpoolOrder, requestOrder } = buildCompatibleOrders(
-                entriesAndRequester,
-                dealTagTee,
+                {
+                    assets: ordersAssets,
+                    requester: requester.address,
+                },
             );
             await signOrder(domain, appOrder, anyone);
 
@@ -828,8 +869,10 @@ describe('IexecPocoBoostDelegate', function () {
             appInstance.owner.returns(appProvider.address);
             datasetInstance.owner.returns(datasetProvider.address);
             const { appOrder, datasetOrder, workerpoolOrder, requestOrder } = buildCompatibleOrders(
-                entriesAndRequester,
-                dealTagTee,
+                {
+                    assets: ordersAssets,
+                    requester: requester.address,
+                },
             );
             await signOrder(domain, appOrder, appProvider);
             await signOrder(domain, datasetOrder, anyone);
@@ -849,8 +892,10 @@ describe('IexecPocoBoostDelegate', function () {
             datasetInstance.owner.returns(datasetProvider.address);
             workerpoolInstance.owner.returns(scheduler.address);
             const { appOrder, datasetOrder, workerpoolOrder, requestOrder } = buildCompatibleOrders(
-                entriesAndRequester,
-                dealTagTee,
+                {
+                    assets: ordersAssets,
+                    requester: requester.address,
+                },
             );
             await signOrder(domain, appOrder, appProvider);
             await signOrder(domain, datasetOrder, datasetProvider);
@@ -871,8 +916,10 @@ describe('IexecPocoBoostDelegate', function () {
             datasetInstance.owner.returns(datasetProvider.address);
             workerpoolInstance.owner.returns(scheduler.address);
             const { appOrder, datasetOrder, workerpoolOrder, requestOrder } = buildCompatibleOrders(
-                entriesAndRequester,
-                dealTagTee,
+                {
+                    assets: ordersAssets,
+                    requester: requester.address,
+                },
             );
             await signOrder(domain, appOrder, appProvider);
             await signOrder(domain, datasetOrder, datasetProvider);
@@ -891,8 +938,11 @@ describe('IexecPocoBoostDelegate', function () {
 
         it('Should fail when the workerpool tag does not provide what app, dataset and request expect', async function () {
             const { appOrder, datasetOrder, workerpoolOrder, requestOrder } = buildCompatibleOrders(
-                entriesAndRequester,
-                dealTagTee,
+                {
+                    assets: ordersAssets,
+                    requester: requester.address,
+                    tag: dealTagTee,
+                },
             );
             // Manually set the tags for app, dataset, and request orders
             appOrder.tag = '0x0000000000000000000000000000000000000000000000000000000000000001'; // 0b0001
@@ -915,8 +965,11 @@ describe('IexecPocoBoostDelegate', function () {
 
         it('Should fail when the last bit of app tag does not provide what dataset or request expect', async function () {
             const { appOrder, datasetOrder, workerpoolOrder, requestOrder } = buildCompatibleOrders(
-                entriesAndRequester,
-                dealTagTee,
+                {
+                    assets: ordersAssets,
+                    requester: requester.address,
+                    tag: dealTagTee,
+                },
             );
 
             // Manually set the tags for app, dataset, and request orders
@@ -941,8 +994,10 @@ describe('IexecPocoBoostDelegate', function () {
 
         it('Should fail when app are different', async function () {
             const { appOrder, datasetOrder, workerpoolOrder, requestOrder } = buildCompatibleOrders(
-                entriesAndRequester,
-                dealTagTee,
+                {
+                    assets: ordersAssets,
+                    requester: requester.address,
+                },
             );
             // Request different app adress
             requestOrder.app = '0x0000000000000000000000000000000000000001';
@@ -959,8 +1014,10 @@ describe('IexecPocoBoostDelegate', function () {
 
         it('Should fail when dataset are different', async function () {
             const { appOrder, datasetOrder, workerpoolOrder, requestOrder } = buildCompatibleOrders(
-                entriesAndRequester,
-                dealTagTee,
+                {
+                    assets: ordersAssets,
+                    requester: requester.address,
+                },
             );
             // Request different dataset adress
             requestOrder.dataset = '0x0000000000000000000000000000000000000001';
@@ -977,8 +1034,10 @@ describe('IexecPocoBoostDelegate', function () {
 
         it('Should fail when requestorder mismatch workerpool restriction ', async function () {
             const { appOrder, datasetOrder, workerpoolOrder, requestOrder } = buildCompatibleOrders(
-                entriesAndRequester,
-                dealTagTee,
+                {
+                    assets: ordersAssets,
+                    requester: requester.address,
+                },
             );
             // Request different dataset adress
             requestOrder.workerpool = '0x0000000000000000000000000000000000000001';
@@ -995,8 +1054,10 @@ describe('IexecPocoBoostDelegate', function () {
 
         it('Should fail when apporder mismatch dataset restriction ', async function () {
             const { appOrder, datasetOrder, workerpoolOrder, requestOrder } = buildCompatibleOrders(
-                entriesAndRequester,
-                dealTagTee,
+                {
+                    assets: ordersAssets,
+                    requester: requester.address,
+                },
             );
             // Request different dataset adress
             appOrder.datasetrestrict = '0x0000000000000000000000000000000000000001';
@@ -1013,8 +1074,10 @@ describe('IexecPocoBoostDelegate', function () {
 
         it('Should fail when apporder mismatch workerpool restriction', async function () {
             const { appOrder, datasetOrder, workerpoolOrder, requestOrder } = buildCompatibleOrders(
-                entriesAndRequester,
-                dealTagTee,
+                {
+                    assets: ordersAssets,
+                    requester: requester.address,
+                },
             );
             // Set different workerpool address
             appOrder.workerpoolrestrict = '0x0000000000000000000000000000000000000001';
@@ -1031,8 +1094,10 @@ describe('IexecPocoBoostDelegate', function () {
 
         it('Should fail when apporder mismatch requester restriction', async function () {
             const { appOrder, datasetOrder, workerpoolOrder, requestOrder } = buildCompatibleOrders(
-                entriesAndRequester,
-                dealTagTee,
+                {
+                    assets: ordersAssets,
+                    requester: requester.address,
+                },
             );
             // Set different requester address
             appOrder.requesterrestrict = '0x0000000000000000000000000000000000000001';
@@ -1049,8 +1114,10 @@ describe('IexecPocoBoostDelegate', function () {
 
         it('Should fail when datasetorder mismatch app restriction', async function () {
             const { appOrder, datasetOrder, workerpoolOrder, requestOrder } = buildCompatibleOrders(
-                entriesAndRequester,
-                dealTagTee,
+                {
+                    assets: ordersAssets,
+                    requester: requester.address,
+                },
             );
             // Set different app address
             datasetOrder.apprestrict = '0x0000000000000000000000000000000000000001';
@@ -1067,8 +1134,10 @@ describe('IexecPocoBoostDelegate', function () {
 
         it('Should fail when datasetorder mismatch workerpool restriction', async function () {
             const { appOrder, datasetOrder, workerpoolOrder, requestOrder } = buildCompatibleOrders(
-                entriesAndRequester,
-                dealTagTee,
+                {
+                    assets: ordersAssets,
+                    requester: requester.address,
+                },
             );
             datasetOrder.workerpoolrestrict = '0x0000000000000000000000000000000000000001';
 
@@ -1084,8 +1153,10 @@ describe('IexecPocoBoostDelegate', function () {
 
         it('Should fail when datasetorder mismatch requester restriction', async function () {
             const { appOrder, datasetOrder, workerpoolOrder, requestOrder } = buildCompatibleOrders(
-                entriesAndRequester,
-                dealTagTee,
+                {
+                    assets: ordersAssets,
+                    requester: requester.address,
+                },
             );
             datasetOrder.requesterrestrict = '0x0000000000000000000000000000000000000001';
 
@@ -1101,8 +1172,10 @@ describe('IexecPocoBoostDelegate', function () {
 
         it('Should fail when workerpoolorder mismatch app restriction', async function () {
             const { appOrder, datasetOrder, workerpoolOrder, requestOrder } = buildCompatibleOrders(
-                entriesAndRequester,
-                dealTagTee,
+                {
+                    assets: ordersAssets,
+                    requester: requester.address,
+                },
             );
             workerpoolOrder.apprestrict = '0x0000000000000000000000000000000000000001';
 
@@ -1118,8 +1191,10 @@ describe('IexecPocoBoostDelegate', function () {
 
         it('Should fail when workerpoolorder mismatch dataset restriction', async function () {
             const { appOrder, datasetOrder, workerpoolOrder, requestOrder } = buildCompatibleOrders(
-                entriesAndRequester,
-                dealTagTee,
+                {
+                    assets: ordersAssets,
+                    requester: requester.address,
+                },
             );
             workerpoolOrder.datasetrestrict = '0x0000000000000000000000000000000000000001';
 
@@ -1135,8 +1210,10 @@ describe('IexecPocoBoostDelegate', function () {
 
         it('Should fail when workerpoolorder mismatch requester restriction', async function () {
             const { appOrder, datasetOrder, workerpoolOrder, requestOrder } = buildCompatibleOrders(
-                entriesAndRequester,
-                dealTagTee,
+                {
+                    assets: ordersAssets,
+                    requester: requester.address,
+                },
             );
             workerpoolOrder.requesterrestrict = '0x0000000000000000000000000000000000000001';
 
@@ -1153,8 +1230,10 @@ describe('IexecPocoBoostDelegate', function () {
         it('Should fail when app not registered', async function () {
             appRegistry.isRegistered.whenCalledWith(appInstance.address).returns(false);
             const { appOrder, datasetOrder, workerpoolOrder, requestOrder } = buildCompatibleOrders(
-                entriesAndRequester,
-                dealTagTee,
+                {
+                    assets: ordersAssets,
+                    requester: requester.address,
+                },
             );
             await signOrder(domain, appOrder, anyone);
 
@@ -1172,7 +1251,10 @@ describe('IexecPocoBoostDelegate', function () {
             appInstance.owner.returns(appProvider.address);
             datasetRegistry.isRegistered.whenCalledWith(datasetInstance.address).returns(false);
             const { orders, appOrder, datasetOrder, workerpoolOrder, requestOrder } =
-                buildCompatibleOrders(entriesAndRequester, dealTagTee);
+                buildCompatibleOrders({
+                    assets: ordersAssets,
+                    requester: requester.address,
+                });
             await signOrders(domain, orders, accounts);
 
             await expect(
@@ -1192,7 +1274,10 @@ describe('IexecPocoBoostDelegate', function () {
                 .whenCalledWith(workerpoolInstance.address)
                 .returns(false);
             const { orders, appOrder, datasetOrder, workerpoolOrder, requestOrder } =
-                buildCompatibleOrders(entriesAndRequester, dealTagTee);
+                buildCompatibleOrders({
+                    assets: ordersAssets,
+                    requester: requester.address,
+                });
             await signOrders(domain, orders, accounts);
 
             await expect(
@@ -1215,13 +1300,11 @@ describe('IexecPocoBoostDelegate', function () {
             const workerpoolPrice = 1_000_000_000;
             const dealPrice = (appPrice + datasetPrice + workerpoolPrice) * volume;
             const { orders, appOrder, datasetOrder, workerpoolOrder, requestOrder } =
-                buildCompatibleOrders(entriesAndRequester, dealTagTee, {
-                    app: appPrice,
-                    dataset: datasetPrice,
-                    workerpool: workerpoolPrice,
+                buildCompatibleOrders({
+                    assets: ordersAssets,
+                    requester: requester.address,
+                    prices: ordersPrices,
                 });
-            requestOrder.requester = requester.address;
-            requestOrder.beneficiary = beneficiary.address;
 
             const initialRequesterBalance = 2;
             await iexecPocoBoostInstance.setVariables({
@@ -1255,13 +1338,11 @@ describe('IexecPocoBoostDelegate', function () {
             const dealPrice = (appPrice + datasetPrice + workerpoolPrice) * volume;
             const schedulerStake = computeSchedulerDealStake(workerpoolPrice, volume);
             const { orders, appOrder, datasetOrder, workerpoolOrder, requestOrder } =
-                buildCompatibleOrders(entriesAndRequester, dealTagTee, {
-                    app: appPrice,
-                    dataset: datasetPrice,
-                    workerpool: workerpoolPrice,
+                buildCompatibleOrders({
+                    assets: ordersAssets,
+                    requester: requester.address,
+                    prices: ordersPrices,
                 });
-            requestOrder.requester = requester.address;
-            requestOrder.beneficiary = beneficiary.address;
 
             const initialRequesterBalance = 2;
             const initialSchedulerBalance = 3; // Way less than what is needed as stake.
@@ -1304,23 +1385,21 @@ describe('IexecPocoBoostDelegate', function () {
 
         it('Should push result (TEE & callback)', async function () {
             workerpoolInstance.m_schedulerRewardRatioPolicy.returns(schedulerRewardRatio);
-            const appPrice = 1000;
-            const datasetPrice = 1_000_000;
-            const workerpoolPrice = 1_000_000_000;
+            const oracleConsumerInstance = await createMock<TestClient__factory, TestClient>(
+                'TestClient',
+            );
             const taskPrice = appPrice + datasetPrice + workerpoolPrice;
             const volume = 3;
             const dealPrice = taskPrice * volume;
             const { orders, appOrder, datasetOrder, workerpoolOrder, requestOrder } =
-                buildCompatibleOrders(
-                    entriesAndRequester,
-                    dealTagTee,
-                    {
-                        app: appPrice,
-                        dataset: datasetPrice,
-                        workerpool: workerpoolPrice,
-                    },
-                    volume,
-                );
+                buildCompatibleOrders({
+                    assets: ordersAssets,
+                    requester: requester.address,
+                    tag: dealTagTee,
+                    prices: ordersPrices,
+                    volume: volume,
+                    callback: oracleConsumerInstance.address,
+                });
             const initialIexecPocoBalance = 1;
             const initialRequesterBalance = 2;
             const initialRequesterFrozen = 3;
@@ -1342,10 +1421,6 @@ describe('IexecPocoBoostDelegate', function () {
                     [requester.address]: initialRequesterFrozen,
                 },
             });
-            const oracleConsumerInstance = await createMock<TestClient__factory, TestClient>(
-                'TestClient',
-            );
-            requestOrder.callback = oracleConsumerInstance.address;
             await signOrders(domain, orders, accounts);
             const dealId = getDealId(domain, requestOrder, taskIndex);
             const taskId = getTaskId(dealId, taskIndex);
@@ -1496,7 +1571,11 @@ describe('IexecPocoBoostDelegate', function () {
 
         it('Should push result (TEE)', async function () {
             const { orders, appOrder, datasetOrder, workerpoolOrder, requestOrder } =
-                buildCompatibleOrders(entriesAndRequester, dealTagTee);
+                buildCompatibleOrders({
+                    assets: ordersAssets,
+                    requester: requester.address,
+                    tag: dealTagTee,
+                });
             await signOrders(domain, orders, accounts);
             const dealId = getDealId(domain, requestOrder, taskIndex);
             const taskId = getTaskId(dealId, taskIndex);
@@ -1536,10 +1615,12 @@ describe('IexecPocoBoostDelegate', function () {
                 .withArgs(dealId, taskIndex, results);
         });
 
-        it('Should push result (Standard)', async function () {
-            const tag = constants.NULL.BYTES32;
+        it('Should push result', async function () {
             const { orders, appOrder, datasetOrder, workerpoolOrder, requestOrder } =
-                buildCompatibleOrders(entriesAndRequester, tag);
+                buildCompatibleOrders({
+                    assets: ordersAssets,
+                    requester: requester.address,
+                });
             await signOrders(domain, orders, accounts);
             const dealId = getDealId(domain, requestOrder, taskIndex);
             const taskId = getTaskId(dealId, taskIndex);
@@ -1576,7 +1657,10 @@ describe('IexecPocoBoostDelegate', function () {
 
         it('Should not push result twice', async function () {
             const { orders, appOrder, datasetOrder, workerpoolOrder, requestOrder } =
-                buildCompatibleOrders(entriesAndRequester, constants.NULL.BYTES32);
+                buildCompatibleOrders({
+                    assets: ordersAssets,
+                    requester: requester.address,
+                });
             await signOrders(domain, orders, accounts);
             const dealId = getDealId(domain, requestOrder, taskIndex);
             const taskId = getTaskId(dealId, taskIndex);
@@ -1614,7 +1698,10 @@ describe('IexecPocoBoostDelegate', function () {
 
         it('Should not push result after deadline', async function () {
             const { orders, appOrder, datasetOrder, workerpoolOrder, requestOrder } =
-                buildCompatibleOrders(entriesAndRequester, dealTagTee);
+                buildCompatibleOrders({
+                    assets: ordersAssets,
+                    requester: requester.address,
+                });
             await signOrders(domain, orders, accounts);
             const startTime = await setNextBlockTimestamp();
             await iexecPocoBoostInstance.matchOrdersBoost(
@@ -1642,7 +1729,11 @@ describe('IexecPocoBoostDelegate', function () {
 
         it('Should not push result without enclave challenge for TEE task', async function () {
             const { orders, appOrder, datasetOrder, workerpoolOrder, requestOrder } =
-                buildCompatibleOrders(entriesAndRequester, dealTagTee);
+                buildCompatibleOrders({
+                    assets: ordersAssets,
+                    requester: requester.address,
+                    tag: dealTagTee,
+                });
             await signOrders(domain, orders, accounts);
             const dealId = getDealId(domain, requestOrder, taskIndex);
             await iexecPocoBoostInstance.matchOrdersBoost(
@@ -1669,7 +1760,10 @@ describe('IexecPocoBoostDelegate', function () {
 
         it('Should not push result with invalid scheduler signature', async function () {
             const { orders, appOrder, datasetOrder, workerpoolOrder, requestOrder } =
-                buildCompatibleOrders(entriesAndRequester, dealTagTee);
+                buildCompatibleOrders({
+                    assets: ordersAssets,
+                    requester: requester.address,
+                });
             await signOrders(domain, orders, accounts);
             await iexecPocoBoostInstance.matchOrdersBoost(
                 appOrder,
@@ -1696,7 +1790,11 @@ describe('IexecPocoBoostDelegate', function () {
 
         it('Should not push result with invalid enclave signature', async function () {
             const { orders, appOrder, datasetOrder, workerpoolOrder, requestOrder } =
-                buildCompatibleOrders(entriesAndRequester, dealTagTee);
+                buildCompatibleOrders({
+                    assets: ordersAssets,
+                    requester: requester.address,
+                    tag: dealTagTee,
+                });
             await signOrders(domain, orders, accounts);
             const dealId = getDealId(domain, requestOrder, taskIndex);
             await iexecPocoBoostInstance.matchOrdersBoost(
@@ -1730,8 +1828,12 @@ describe('IexecPocoBoostDelegate', function () {
 
         it('Should not push result with missing data for callback', async function () {
             const { orders, appOrder, datasetOrder, workerpoolOrder, requestOrder } =
-                buildCompatibleOrders(entriesAndRequester, dealTagTee);
-            requestOrder.callback = '0x000000000000000000000000000000000000ca11';
+                buildCompatibleOrders({
+                    assets: ordersAssets,
+                    requester: requester.address,
+                    tag: dealTagTee,
+                    callback: ethers.Wallet.createRandom().address,
+                });
             await signOrders(domain, orders, accounts);
             const dealId = getDealId(domain, requestOrder, taskIndex);
             const taskId = getTaskId(dealId, taskIndex);
@@ -1782,23 +1884,16 @@ describe('IexecPocoBoostDelegate', function () {
         });
 
         it('Should claim', async function () {
-            const appPrice = 1000;
-            const datasetPrice = 1_000_000;
-            const workerpoolPrice = 1_000_000_000;
             const expectedVolume = 2; // > 1 to explicit taskPrice vs dealPrice
             const taskPrice = appPrice + datasetPrice + workerpoolPrice;
             const dealPrice = taskPrice * expectedVolume;
             const { orders, appOrder, datasetOrder, workerpoolOrder, requestOrder } =
-                buildCompatibleOrders(
-                    entriesAndRequester,
-                    dealTagTee,
-                    {
-                        app: appPrice,
-                        dataset: datasetPrice,
-                        workerpool: workerpoolPrice,
-                    },
-                    expectedVolume,
-                );
+                buildCompatibleOrders({
+                    assets: ordersAssets,
+                    requester: requester.address,
+                    prices: ordersPrices,
+                    volume: expectedVolume,
+                });
             await signOrders(domain, orders, accounts);
             const initialIexecPocoBalance = 1;
             const initialRequesterBalance = 2;
@@ -1874,7 +1969,10 @@ describe('IexecPocoBoostDelegate', function () {
 
         it('Should not claim if task not unset', async function () {
             const { orders, appOrder, datasetOrder, workerpoolOrder, requestOrder } =
-                buildCompatibleOrders(entriesAndRequester, standardDealTag);
+                buildCompatibleOrders({
+                    assets: ordersAssets,
+                    requester: requester.address,
+                });
             await signOrders(domain, orders, accounts);
             const dealId = getDealId(domain, requestOrder, taskIndex);
             const taskId = getTaskId(dealId, taskIndex);
@@ -1920,7 +2018,10 @@ describe('IexecPocoBoostDelegate', function () {
 
         it('Should not claim if task unknown because of wrong index', async function () {
             const { orders, appOrder, datasetOrder, workerpoolOrder, requestOrder } =
-                buildCompatibleOrders(entriesAndRequester, standardDealTag);
+                buildCompatibleOrders({
+                    assets: ordersAssets,
+                    requester: requester.address,
+                });
             await signOrders(domain, orders, accounts);
             const dealId = getDealId(domain, requestOrder, taskIndex);
             await iexecPocoBoostInstance.matchOrdersBoost(
@@ -1940,7 +2041,10 @@ describe('IexecPocoBoostDelegate', function () {
 
         it('Should not claim before deadline', async function () {
             const { orders, appOrder, datasetOrder, workerpoolOrder, requestOrder } =
-                buildCompatibleOrders(entriesAndRequester, standardDealTag);
+                buildCompatibleOrders({
+                    assets: ordersAssets,
+                    requester: requester.address,
+                });
             await signOrders(domain, orders, accounts);
             const dealId = getDealId(domain, requestOrder, taskIndex);
             const startTime = await setNextBlockTimestamp();